Make /admin/coaches HTML index action require admin
Only the JSON format doesn't require admin since it's loading
from the home page for every user.

This probably isn't the cleanest solution, but as of right now,
we don't really need a separate controller just for that one action.
It should be added in the future to the API.